Change StepIndex to be signed
I have a place where I want it to be able to be -1. Besides,
unsigned arithmetic is scary.

+++ src/shk/src/manifest/step.h
@@ -14,7 +14,7 @@ namespace shk {
  * that vector, or into a vector of the same length that refers to the same Step
  * objects.
  */
-using StepIndex = size_t;
+using StepIndex = int;
 
 /**
  * A Step is a dumb data object that represents one build statment in the
